[Fix] Fix chat templating in Mini-SWE-Agent and Terminal-Bench examples (#404)
Fixes chat templating in the Mini-swe-agent and the terminal bench
examples.

Previously, we were naively callling `.apply_chat_template` to encode
response messages turn by turn - but this can append system messages for
each turn depending on the model. (h/t to @CharlieFRuan )

For example, Qwen 3 8B , the templating works fine, but then for Qwen
2.5 1.5B instruct, the code adds a system prompt message while
tokenizing every turn

We use the fixed base approach similar to what we do in the
SkyRLGymGenerator.

---------

Signed-off-by: SumanthRH <sumanthrh99@gmail.com>